# Conflicts:
#	.gitignore
#	doc/source/api.rst.mako
#	pyramid_oereb/lib/records/documents.py
#	pyramid_oereb/lib/records/geometry.py
#	pyramid_oereb/lib/records/real_estate.py
#	pyramid_oereb/lib/records/reference_definition.py
#	pyramid_oereb/lib/records/view_service.py
#	pyramid_oereb/lib/sources/__init__.py
#	pyramid_oereb/lib/sources/plr.py

+++ pyramid_oereb/standard/__init__.py
@@ -6,6 +6,7 @@ import re
 from mako.template import Template
 from pyramid.path import AssetResolver, DottedNameResolver
 from sqlalchemy import create_engine
+from sqlalchemy.schema import CreateTable
 from shutil import copyfile
 from pyramid_oereb.lib.config import parse
 
@@ -20,7 +21,7 @@ def convert_camel_case_to_text_form(name):
     return re.sub('([a-z0-9])([A-Z])', r'\1 \2', s1)
 
 
-def _create_standard_configuration_models_py_(code, geometry_type, absolute_path):
+def _create_standard_configuration_models_py_(code, geometry_type, absolute_path, schema=None):
     """
     The simplest way to get a python file containing a database definition in sqlalchemy orm way. It will
      contain all necessary definitions to produce an extract as the specification defines for the new topic.
@@ -30,6 +31,7 @@ def _create_standard_configuration_models_py_(code, geometry_type, absolute_path
         geometry_type (str): A valid geometry type.
         absolute_path (str): The absolute Path where the genderated python file will be placed. It
             must bewriteable by the user running this command.
+        schema (str): The schema name. If not specified, "name" will be used.
     """
     template = Template(
         filename=AssetResolver('pyramid_oereb').resolve('standard/templates/plr.py.mako').abspath()
@@ -37,7 +39,7 @@ def _create_standard_configuration_models_py_(code, geometry_type, absolute_path
     name = convert_camel_case_to_snake_case(code)
     content = template.render(**{
         'topic': convert_camel_case_to_text_form(code),
-        'schema_name': name,
+        'schema_name': schema or name,
         'geometry_type': geometry_type
     })
     models_path = '{path}/{name}.py'.format(
@@ -62,8 +64,7 @@ def _create_all_standard_models_by_yaml_(configuration_yaml_path, section='pyram
     Args:
         configuration_yaml_path (str): The absolute path to the yaml file which contains the plr
             definitions.
-        section (str): The section in yaml file where the plrs are configured in. Standard is
-            'pyramid
+        section (str): The section in yaml file where the plrs are configured in. Default is 'pyramid_oereb'.
     """
     config = parse(configuration_yaml_path, section)
     absolute_path = AssetResolver('pyramid_oereb.standard.models').resolve('').abspath()
@@ -76,7 +77,8 @@ def _create_all_standard_models_by_yaml_(configuration_yaml_path, section='pyram
             )
 
 
-def _create_tables_from_standard_configuration_(configuration_yaml_path, section='pyramid_oereb'):
+def create_tables_from_standard_configuration(
+        configuration_yaml_path, section='pyramid_oereb', tables_only=False, sql_file=None):
     """
     Creates all schemas which are defined in the passed yaml file: <section>.<plrs>.[<plr>.<code>]. The code
     must be camel case. It will be transformed to snake case and used as schema name.
@@ -86,31 +88,64 @@ def _create_tables_from_standard_configuration_(configuration_yaml_path, section
     Args:
         configuration_yaml_path (str): The absolute path to the yaml file which contains the plr
             definitions.
-        section (str): The section in yaml file where the plrs are configured in. Standard is
-            'pyramid
+        section (str): The section in yaml file where the plrs are configured in. Default is 'pyramid_oereb'.
+        tables_only (bool): True to skip creation of schema. Default is False.
+        sql_file (file): the file to generate. Default is None (in the database).
     """
     config = parse(configuration_yaml_path, section)
-    main_schema_engine = create_engine(config.get('app_schema').get('db_connection'), echo=True)
-    main_schema_connection = main_schema_engine.connect()
-    main_schema_connection.execute('CREATE SCHEMA {name};'.format(name=config.get('app_schema').get('name')))
+
+    if sql_file is None:
+        main_schema_engine = create_engine(config.get('app_schema').get('db_connection'), echo=True)
+        if not tables_only:
+            main_schema_connection = main_schema_engine.connect()
+            try:
+                main_schema_connection.execute(
+                    'CREATE SCHEMA IF NOT EXISTS {name};'.format(name=config.get('app_schema').get('name'))
+                )
+            finally:
+                main_schema_connection.close()
+    else:
+        sql_file.write('CREATE SCHEMA {name};\n'.format(name=config.get('app_schema').get('name')))
+
     main_base_class = DottedNameResolver().maybe_resolve('{package}.Base'.format(
         package=config.get('app_schema').get('models')
     ))
-    main_base_class.metadata.create_all(main_schema_engine)
+    if sql_file is None:
+        main_base_class.metadata.create_all(main_schema_engine)
+    else:
+        for table in main_base_class.metadata.sorted_tables:
+            sql_file.write('{};\n'.format(str(CreateTable(table)).replace('DATETIME', 'timestamp')))
     for schema in config.get('plrs'):
-        plr_schema_engine = create_engine(schema.get('source').get('params').get('db_connection'), echo=True)
-        plr_schema_connection = plr_schema_engine.connect()
-        plr_schema_connection.execute('CREATE SCHEMA {name};'.format(
-            name=convert_camel_case_to_snake_case(schema.get('code')))
-        )
-        plr_base = DottedNameResolver().maybe_resolve('{package}.Base'.format(
-            package=schema.get('source').get('params').get('models')
-        ))
-        plr_schema_connection.close()
-        plr_base.metadata.create_all(plr_schema_engine)
+        if sql_file is None:
+            if schema.get('standard'):
+                plr_schema_engine = create_engine(schema.get('source').get('params').get('db_connection'),
+                                                  echo=True)
+
+                if not tables_only:
+                    plr_schema_connection = plr_schema_engine.connect()
+                    try:
+                        plr_schema_connection.execute('CREATE SCHEMA IF NOT EXISTS {name};'.format(
+                            name=convert_camel_case_to_snake_case(schema.get('code')))
+                        )
+                    finally:
+                        plr_schema_connection.close()
+
+                plr_base = DottedNameResolver().maybe_resolve('{package}.Base'.format(
+                    package=schema.get('source').get('params').get('models')
+                ))
+                plr_base.metadata.create_all(plr_schema_engine)
+        else:
+            plr_base = DottedNameResolver().maybe_resolve('{package}.Base'.format(
+                package=schema.get('source').get('params').get('models')
+            ))
+            sql_file.write('CREATE SCHEMA {name};\n'.format(
+                name=convert_camel_case_to_snake_case(schema.get('code')))
+            )
+            for table in plr_base.metadata.sorted_tables:
+                sql_file.write('{};\n'.format(str(CreateTable(table)).replace('DATETIME', 'timestamp')))
 
 
-def _drop_tables_from_standard_configuration_(configuration_yaml_path, section='pyramid_oereb'):
+def drop_tables_from_standard_configuration(configuration_yaml_path, section='pyramid_oereb'):
     """
     Drops all schemas which are defined in the passed yaml file: <section>.<plrs>.[<plr>.<code>]. The code
     must be camel case. It will be transformed to snake case and used as schema name.
@@ -119,8 +154,7 @@ def _drop_tables_from_standard_configuration_(configuration_yaml_path, section='
     Args:
         configuration_yaml_path (str): The absolute path to the yaml file which contains the plr
             definitions.
-        section (str): The section in yaml file where the plrs are configured in. Standard is
-            'pyramid
+        section (str): The section in yaml file where the plrs are configured in. Default is 'pyramid_oereb'.
     """
     config = parse(configuration_yaml_path, section)
     main_schema_engine = create_engine(config.get('app_schema').get('db_connection'), echo=True)
@@ -130,12 +164,14 @@ def _drop_tables_from_standard_configuration_(configuration_yaml_path, section='
     )
     main_schema_connection.close()
     for schema in config.get('plrs'):
-        plr_schema_engine = create_engine(schema.get('source').get('params').get('db_connection'), echo=True)
-        plr_schema_connection = plr_schema_engine.connect()
-        plr_schema_connection.execute('DROP SCHEMA IF EXISTS {name} CASCADE;'.format(
-            name=convert_camel_case_to_snake_case(schema.get('code')))
-        )
-        plr_schema_connection.close()
+        if schema.get('standard'):
+            plr_schema_engine = create_engine(schema.get('source').get('params').get('db_connection'),
+                                              echo=True)
+            plr_schema_connection = plr_schema_engine.connect()
+            plr_schema_connection.execute('DROP SCHEMA IF EXISTS {name} CASCADE;'.format(
+                name=convert_camel_case_to_snake_case(schema.get('code')))
+            )
+            plr_schema_connection.close()
 
 
 def _create_standard_yaml_config_(name='pyramid_oereb_standard.yml',
@@ -148,7 +184,6 @@ def _create_standard_yaml_config_(name='pyramid_oereb_standard.yml',
         (str): The name of the new file. Default
         database (str): The database connection string.Default:
             'postgresql://postgres:password@localhost/pyramid_oereb'
-
     """
 
     # File names
